добавил слой абстракции на сервис, + мелкие правки
