+++ spring-07/pom.xml
@@ -14,7 +14,7 @@
     <name>spring-07</name>
 
     <properties>
-        <java.version>8</java.version>
+        <java.version>11</java.version>
     </properties>
 
     <dependencies>
