+++ spring-07/src/test/java/ru/otus/spring/dao/BookDaoJdbcTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import org.springframework.test.annotation.DirtiesContext;
 import ru.otus.spring.domain.Author;
 import ru.otus.spring.domain.Book;
 import ru.otus.spring.domain.Jenre;
@@ -12,8 +11,6 @@ import ru.otus.spring.domain.Jenre;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.test.annotation.DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD;
-import static ru.otus.spring.domain.Jenre.*;
 
 @JdbcTest
 @Import(BookDaoJdbc.class)
